modified:   vllm/model_executor/layers/layernorm.py

+++ vllm/model_executor/layers/layernorm.py
@@ -12,7 +12,7 @@ import triton.language as tl
 def rms_norm_kernel(x, weight, epsilon, output, residual, num_elements):
     pid = tl.program_id(0)
     block_size = 64
-    idx = pid * block_size + tl.arange(0, block_size, dtype=tl.int32)
+    idx = pid * block_size + tl.arange(0, block_size)
 
     # 使用掩码避免越界
     mask = idx < num_elements
